src/requirements.txt
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1
pydub==0.25.1
simpleaudio==1.0.4
